# Improve parsing speed by about a factor of 2 for large files