This PR has three improvements to the parser, one is avoiding reallocating the line buffer over and over again, one is to use a lookup table to avoid expensive calls to pow(), and the final one avoids the ldexp() and pow() calls for the case that the exponent is zero.

Parse times for some large .obj files (without asan):

```
        File A  File B  File C
Before  2743ms   589ms   615ms
After   1037ms   203ms   190ms
```