@dotdash 2016-10-24
After seeing PR #65 I tested this with the `powerplant.obj` file mentioned there, and I get down from 9.8s to 7.5s. The impact is smaller because that files has less digits past the decimal point.