@dotdash 2016-10-24
With the third commit, the power plant test is down to 7s :-)