@syoyo 2016-10-24
Super cool! merged.

I have to update the performance table for RungHolt scene in README.md :-)