@syoyo 2016-10-24
FYI, `rungholt.obj` scene does not gain the performance from this PR(actually the performance gets a bit worser. I got 7.1 secs on recent `master`, while previously it was 6.8 secs according to `README.md` log).

Anyway, I'm not so serious about this result. As you mentioned, the performance depends on file content > The impact is smaller because that files has less digits past the decimal point.