@dotdash 2016-10-24
Hm, that's curious, I'm getting down from 3.65s to 2.79s. Measured using `perf stat -r5 ./loader_example rungholt.obj`. Compiled without asan and the `PrintInfo` line is commented out.

With asan, it's still down from ~10s to ~9.2s, but I only did a single run of each in that setup.

What hardware/OS did you test that on?