@dotdash 2016-10-24
Oh, FWIW, I'm running on an i7 3770 on a Debian Linux