@syoyo 2016-10-24
> What hardware/OS did you test that on?

As described in `README.md`, I'm using MacBook 12(and macOS Sierra). And I'm using `viewer` example to measure parsing time.

> I'm running on an i7 3770 on a Debian Linux

I guess Desktop CPU does good job for the performance gain :-)